python/dns-service/requirements.txt
attrs==24.2.0
black==24.8.0
cattrs==24.1.2
click==8.1.7
docstring-to-markdown==0.15
grpcio==1.66.2
grpcio-tools==1.66.2
isort==5.13.2
jedi==0.19.1
jedi-language-server==0.41.4
lsprotocol==2023.0.1
mypy-extensions==1.0.0
packaging==24.1
parso==0.8.4
pathspec==0.12.1
pip==24.2
platformdirs==4.3.6
pluggy==1.5.0
protobuf==5.28.2
pygls==1.3.1
python-lsp-jsonrpc==1.1.2
python-lsp-server==1.12.0
ruff==0.6.9
setuptools==75.1.0
ujson==5.10.0
